requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
av==10.0.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
colorama==0.4.6
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
daphne==4.0.0
Django==3.2.20
django-extensions==3.2.3
djangorestframework==3.14.0
filelock==3.12.2
fonttools==4.40.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scipy==1.11.1
seaborn==0.12.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.132
urllib3==2.0.3
zope.interface==6.0